@@ -1,129 +1,98 @@
-import { Action } from '@ngrx/store';
+import { createAction, props } from '@ngrx/store';
 import { MatrixData, Task, Topic } from './matrix.interfaces';
 
-export const GET_MATRIX_DATA = '[MATRIX] get all matrix data from sever';
-export const GET_MATRIX_DATA_SUCCESS =
-  '[MATRIX] successfully got all matrix data';
-export const GET_MATRIX_DATA_FAILED = '[MATRIX] failed to get all matrix data';
+export const getMatrixData = createAction(
+  '[MATRIX] get all matrix data from sever',
+);
 
-export const UPDATE_TASK = '[MATRIX] update this task';
-export const UPDATE_TASK_SUCCESS = '[MATRIX] successfully updated task';
-export const UPDATE_TASK_FAILED = '[MATRIX] failed to update task';
+export const getMatrixDataSuccess = createAction(
+  '[MATRIX] successfully got all matrix data',
+  props<{ data: MatrixData }>(),
+);
 
-export const DELETE_TASK = '[MATRIX] delete this task';
-export const DELETE_TASK_SUCCESS = '[MATRIX] successfully deleted task';
-export const DELETE_TASK_FAILED = '[MATRIX] failed to delete task';
+export const getMatrixDataFailed = createAction(
+  '[MATRIX] failed to get all matrix data',
+  props<{ message: string }>(),
+);
 
-export const ADD_TOPIC = '[MATRIX] add new topic';
-export const ADD_TOPIC_SUCCESS = '[MATRIX] successfully added topic';
-export const ADD_TOPIC_FAILED = '[MATRIX] failed to add topic';
+export const updateTask = createAction(
+  '[MATRIX] update this task',
+  props<{ task: Task }>(),
+);
 
-export const UPDATE_TOPIC = '[MATRIX] update this topic';
-export const UPDATE_TOPIC_SUCCESS = '[MATRIX] successfully updated topic';
-export const UPDATE_TOPIC_FAILED = '[MATRIX] failed to update topic';
+export const updateTaskSuccess = createAction(
+  '[MATRIX] successfully updated task',
+  props<{ task: Task }>(),
+);
 
-export const SELECT_TASK = '[MATRIX] select a task to why details';
-export const TOGGLE_TOPIC_VISIBLITY =
-  '[MATRIX] toggle topic visibility on or off';
+export const updateTaskFailed = createAction(
+  '[MATRIX] failed to update task',
+  props<{ message: string }>(),
+);
 
-export class GetMatrixData implements Action {
-  readonly type = GET_MATRIX_DATA;
-}
+export const deleteTask = createAction(
+  '[MATRIX] delete this task',
+  props<{ taskId: number }>(),
+);
 
-export class GetMatrixDataSuccess implements Action {
-  readonly type = GET_MATRIX_DATA_SUCCESS;
-  constructor(public data: MatrixData) {}
-}
+export const deleteTaskSuccess = createAction(
+  '[MATRIX] successfully deleted task',
+  props<{ task: Task }>(),
+);
 
-export class GetMatrixDataFailed implements Action {
-  readonly type = GET_MATRIX_DATA_FAILED;
-  constructor(public message: string) {}
-}
+export const deleteTaskFailed = createAction(
+  '[MATRIX] failed to delete task',
+  props<{ message: string }>(),
+);
 
-export class UpdateTask implements Action {
-  constructor(public task: Task) {}
-  readonly type = UPDATE_TASK;
-}
+export const toggleDoneTask = createAction(
+  '[MATRIX] toggle task done',
+  props<{ taskId: number }>(),
+);
 
-export class UpdateTaskSuccess implements Action {
-  readonly type = UPDATE_TASK_SUCCESS;
-  constructor(public task: Task) {}
-}
+export const toggleDoneTaskSuccess = createAction(
+  '[MATRIX] successfully toggled task done',
+  props<{ task: Task }>(),
+);
 
-export class UpdateTaskFailed implements Action {
-  readonly type = UPDATE_TASK_FAILED;
-  constructor(public message: string) {}
-}
+export const toggleDoneTaskFailed = createAction(
+  '[MATRIX] failed to toggle task done',
+  props<{ message: string }>(),
+);
 
-export class DeleteTask implements Action {
-  readonly type = DELETE_TASK;
-  constructor(public taskId: number) {}
-}
+export const addTopic = createAction('[MATRIX] add new topic');
 
-export class DeleteTaskSuccess implements Action {
-  readonly type = DELETE_TASK_SUCCESS;
-  constructor(public task: Task) {}
-}
+export const addTopicSuccess = createAction(
+  '[MATRIX] successfully added topic',
+  props<{ topic: Topic }>(),
+);
 
-export class DeleteTaskFailed implements Action {
-  readonly type = DELETE_TASK_FAILED;
-  constructor(public message: string) {}
-}
+export const addTopicFailed = createAction(
+  '[MATRIX] failed to add topic',
+  props<{ message: string }>(),
+);
 
-export class AddTopic implements Action {
-  readonly type = ADD_TOPIC;
-}
+export const updateTopic = createAction(
+  '[MATRIX] update this topic',
+  props<{ topic: Topic }>(),
+);
 
-export class AddTopicSuccess implements Action {
-  readonly type = ADD_TOPIC_SUCCESS;
-  constructor(public topic: Topic) {}
-}
+export const updateTopicSuccess = createAction(
+  '[MATRIX] successfully updated topic',
+  props<{ topic: Topic }>(),
+);
 
-export class AddTopicFailed implements Action {
-  readonly type = ADD_TOPIC_FAILED;
-  constructor(public message: string) {}
-}
+export const updateTopicFailed = createAction(
+  '[MATRIX] failed to update topic',
+  props<{ message: string }>(),
+);
 
-export class UpdateTopic implements Action {
-  readonly type = UPDATE_TOPIC;
-  constructor(public topic: Topic) {}
-}
+export const selectTask = createAction(
+  '[MATRIX] select a task to why details',
+  props<{ currentTaskId: number }>(),
+);
 
-export class UpdateTopicSuccess implements Action {
-  readonly type = UPDATE_TOPIC_SUCCESS;
-  constructor(public topic: Topic) {}
-}
-
-export class UpdateTopicFailed implements Action {
-  readonly type = UPDATE_TOPIC_FAILED;
-  constructor(public message: string) {}
-}
-
-export class SelectTask implements Action {
-  readonly type = SELECT_TASK;
-  constructor(public currentTaskId: number) {}
-}
-
-export class ToggleTopicVisibility implements Action {
-  readonly type = TOGGLE_TOPIC_VISIBLITY;
-  constructor(public topicId: number) {}
-}
-
-export type MatrixAction =
-  | GetMatrixData
-  | GetMatrixDataSuccess
-  | GetMatrixDataFailed
-  | UpdateTask
-  | UpdateTaskSuccess
-  | UpdateTaskFailed
-  | DeleteTask
-  | DeleteTaskSuccess
-  | DeleteTaskFailed
-  | AddTopic
-  | AddTopicSuccess
-  | AddTopicFailed
-  | UpdateTopic
-  | UpdateTopicSuccess
-  | UpdateTopicFailed
-  | SelectTask
-  | ToggleTopicVisibility;
+export const toggleTopicVisibility = createAction(
+  '[MATRIX] toggle topic visibility on or off',
+  props<{ topicId: number }>(),
+);
