@@ -1,8 +1,17 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import { Store } from '@ngrx/store';
+import { AppState } from '../app.state';
+import { getMatrixData } from './matrix.actions';
 
 @Component({
   selector: 'app-matrix',
   templateUrl: './matrix.component.html',
   styleUrls: ['./matrix.component.scss'],
 })
-export class MatrixComponent {}
+export class MatrixComponent implements OnInit {
+  constructor(private readonly store: Store<AppState>) {}
+
+  ngOnInit(): void {
+    this.store.dispatch(getMatrixData());
+  }
+}
