@@ -1,36 +1,29 @@
 import { Observable } from 'rxjs';
 import { Color } from '../shared/color.interfaces';
+import { RequestStatus } from '../shared/request-status.interface';
 
-export class Topic {
-  constructor(
-    public id: number,
-    public name: string,
-    public color: Color,
-    public visible: boolean,
-    public deleted: boolean,
-  ) {}
-
-  toggleVisibility() {
-    this.visible = !this.visible;
-  }
+export interface Topic {
+  id: number;
+  name: string;
+  color: Color;
+  visible: boolean;
+  deleted: boolean;
 }
 
-export class Task {
-  constructor(
-    public id: number,
-    public name: string,
-    public topic: number,
-    public importance: number,
-    public dueDay: number,
-    public dueMonth: number,
-    public dueYear: number,
-    public description: string = '',
-    public done: boolean = false,
-    public deleted: boolean = false,
-    public createdAt: Date = new Date(),
-    public updatedAt: Date = new Date(),
-    public deletedAt: Date = new Date(),
-  ) {}
+export interface Task {
+  id: number;
+  name: string;
+  importance: number;
+  topic: number;
+  dueDay: number;
+  dueMonth: number;
+  dueYear: number;
+  description?: string;
+  done: boolean;
+  deleted: boolean;
+  createdAt: Date;
+  updatedAt?: Date;
+  deletedAt?: Date;
 }
 
 export interface TaskDictionary {
@@ -46,26 +39,22 @@ export interface MatrixData {
   tasks: Task[];
 }
 
-export interface MatrixState {
-  topics: Topic[];
-  tasks: Task[];
-  taskHistory: number[];
-  isLoading: boolean;
-  errorMessage?: string;
+export interface DropdownItem {
+  id: number;
+  value: string;
+  color: string;
 }
 
 export interface MatrixService {
-  getData: () => void;
-  updateTask: (task: Task) => void;
-  deleteTask: (taskId: number) => void;
-  selectTask: (task: Task) => void;
-  toggleTopicVisibility: (topicId: number) => void;
-  updateTopic: (topic: Topic) => void;
   selectTopics: () => Observable<Topic[]>;
-  selectIsLoading: () => Observable<boolean>;
-  selectErrorMessage: () => Observable<string | undefined>;
+  selectRequestStatus: () => Observable<RequestStatus>;
   selectTasks: () => Observable<Task[]>;
+  selectActiveTasks: () => Observable<Task[]>;
+  selectDoneTasks: () => Observable<Task[]>;
   selectTasksByTopics: () => Observable<TaskDictionary>;
-  selectTaskHistory: () => Observable<Task[]>;
+  selectActiveTasksByTopics: () => Observable<TaskDictionary>;
+  selectDoneTasksByTopics: () => Observable<TaskDictionary>;
+  selectCurrentTaskHistory: () => Observable<Task[]>;
+  selectMatrixTaskHistory: () => Observable<number[]>;
   selectTopicById: (id: number) => Observable<Topic | undefined>;
 }
