@@ -1,111 +1,133 @@
-import { Action } from '@ngrx/store';
-import produce from 'immer';
-import {
-  ADD_TOPIC,
-  ADD_TOPIC_FAILED,
-  ADD_TOPIC_SUCCESS,
-  DELETE_TASK,
-  DELETE_TASK_FAILED,
-  DELETE_TASK_SUCCESS,
-  GET_MATRIX_DATA,
-  GET_MATRIX_DATA_FAILED,
-  GET_MATRIX_DATA_SUCCESS,
-  MatrixAction,
-  SELECT_TASK,
-  TOGGLE_TOPIC_VISIBLITY,
-  UPDATE_TASK,
-  UPDATE_TASK_FAILED,
-  UPDATE_TASK_SUCCESS,
-  UPDATE_TOPIC,
-  UPDATE_TOPIC_FAILED,
-  UPDATE_TOPIC_SUCCESS,
-} from './matrix.actions';
-import { MatrixState } from './matrix.interfaces';
+import { Action, createReducer, on } from '@ngrx/store';
+import { cloneDeep, findIndex, pull } from 'lodash';
+import { RequestStatus } from '../shared/request-status.interface';
+import * as MatrixActions from './matrix.actions';
+import { Task, Topic } from './matrix.interfaces';
+
+export interface MatrixState {
+  topics: Topic[];
+  tasks: Task[];
+  taskHistory: number[];
+  requestStatus: RequestStatus;
+}
+
+export const matrixFeatureKey = 'matrix';
 
 export const initialState: MatrixState = {
   topics: [],
   tasks: [],
   taskHistory: [],
-  isLoading: false,
-  errorMessage: undefined,
+  requestStatus: {
+    isLoading: false,
+    errorMessage: undefined,
+  },
 };
 
-export function matrixReducer(state = initialState, action: Action) {
-  return produce((draft: MatrixState, matrixAction: MatrixAction) => {
-    switch (matrixAction.type) {
-      // === loading cases ===
-      case GET_MATRIX_DATA:
-      case UPDATE_TASK:
-      case UPDATE_TOPIC:
-      case ADD_TOPIC:
-      case DELETE_TASK:
-        draft.isLoading = true;
-        draft.errorMessage = undefined;
-        return;
-      // === failed cases ===
-      case GET_MATRIX_DATA_FAILED:
-      case UPDATE_TASK_FAILED:
-      case UPDATE_TOPIC_FAILED:
-      case ADD_TOPIC_FAILED:
-      case DELETE_TASK_FAILED:
-        draft.isLoading = false;
-        draft.errorMessage = matrixAction.message;
-        return;
-      // === success cases ===
-      case GET_MATRIX_DATA_SUCCESS:
-        draft.topics = matrixAction.data.topics;
-        draft.tasks = matrixAction.data.tasks;
-        setUnloading(draft);
-        return;
-      case UPDATE_TASK_SUCCESS:
-      case DELETE_TASK_SUCCESS:
-        const taskIndex = draft.tasks.findIndex(
-          (task) => task.id === matrixAction.task.id,
-        );
-        if (taskIndex > -1) {
-          draft.tasks[taskIndex] = matrixAction.task;
-        }
-        setUnloading(draft);
-        return;
-      case ADD_TOPIC_SUCCESS:
-        draft.topics.push(matrixAction.topic);
-        setUnloading(draft);
-        return;
-      case UPDATE_TOPIC_SUCCESS:
-        const topicIndex = draft.topics.findIndex(
-          (topic) => topic.id === matrixAction.topic.id,
-        );
-        if (topicIndex > -1) {
-          draft.topics[topicIndex] = matrixAction.topic;
-        }
-        setUnloading(draft);
-        return;
-      // === frontend only actions ===
-      case SELECT_TASK:
-        const alreadyInHistory = draft.taskHistory.findIndex(
-          (t) => t === matrixAction.currentTaskId,
-        );
-        if (alreadyInHistory > -1) {
-          draft.taskHistory.splice(alreadyInHistory, 1);
-        }
-        draft.taskHistory.unshift(matrixAction.currentTaskId);
-        if (draft.taskHistory.length > 6) {
-          draft.taskHistory.pop();
-        }
-        return;
-      case TOGGLE_TOPIC_VISIBLITY:
-        const topicIdx = draft.topics.findIndex(
-          (topic) => topic.id === matrixAction.topicId,
-        );
-        if (topicIdx > -1) {
-          draft.topics[topicIdx].toggleVisibility();
-        }
-        return;
-    }
-  })(state, action as MatrixAction);
+export function matrixReducer(
+  state: MatrixState | undefined = initialState,
+  action: Action,
+): MatrixState {
+  return {
+    topics: topicsReducer(state.topics, action),
+    tasks: tasksReducer(state.tasks, action),
+    taskHistory: taskHistoryReducer(state.taskHistory, action),
+    requestStatus: requestStatusReducer(state.requestStatus, action),
+  };
 }
 
-const setUnloading = (draft: MatrixState): void => {
-  draft.isLoading = false;
-  draft.errorMessage = undefined;
-};
+const topicsReducer = createReducer(
+  initialState.topics,
+  on(MatrixActions.getMatrixDataSuccess, (state, { data }) => data.topics),
+  on(MatrixActions.addTopicSuccess, (state, { topic }) => [
+    ...cloneDeep(state),
+    topic,
+  ]),
+  on(MatrixActions.updateTopicSuccess, (state, { topic }) => {
+    const topicIndex = findIndex(state, { id: topic.id });
+    if (topicIndex > -1) {
+      var newTopics = [...cloneDeep(state)];
+      newTopics[topicIndex] = topic;
+      return newTopics;
+    }
+    return state;
+  }),
+  on(MatrixActions.toggleTopicVisibility, (state, { topicId }) => {
+    const topicIdx = findIndex(state, { id: topicId });
+    if (topicIdx > -1) {
+      var toggleTopics = [...cloneDeep(state)];
+      toggleTopics[topicIdx].visible = !toggleTopics[topicIdx].visible;
+      return toggleTopics;
+    }
+    return state;
+  }),
+);
+
+const tasksReducer = createReducer(
+  initialState.tasks,
+  on(MatrixActions.getMatrixDataSuccess, (state, { data }) => data.tasks),
+  on(
+    MatrixActions.updateTaskSuccess,
+    MatrixActions.deleteTaskSuccess,
+    (state, { task }) => {
+      const taskIdx = findIndex(state, { id: task.id });
+      if (taskIdx > -1) {
+        var newTasks = [...cloneDeep(state)];
+        newTasks[taskIdx] = task;
+        return newTasks;
+      }
+      return state;
+    },
+  ),
+  on(MatrixActions.toggleDoneTask, (state, { taskId }) => {
+    const taskIdx = findIndex(state, { id: taskId });
+    if (taskIdx > -1) {
+      var toggleTasks = [...cloneDeep(state)];
+      toggleTasks[taskIdx].done = !toggleTasks[taskIdx].done;
+      return toggleTasks;
+    }
+    return state;
+  }),
+);
+
+const taskHistoryReducer = createReducer(
+  initialState.taskHistory,
+  on(MatrixActions.deleteTaskSuccess, (state, { task }) =>
+    pull(cloneDeep(state), task.id),
+  ),
+  on(MatrixActions.selectTask, (state, { currentTaskId }) => {
+    var newHistory = pull(cloneDeep(state), currentTaskId);
+    newHistory.unshift(currentTaskId);
+    if (newHistory.length > 6) {
+      newHistory.pop();
+    }
+    return newHistory;
+  }),
+);
+
+const requestStatusReducer = createReducer(
+  initialState.requestStatus,
+  on(
+    MatrixActions.getMatrixData,
+    MatrixActions.updateTask,
+    MatrixActions.deleteTask,
+    MatrixActions.updateTopic,
+    MatrixActions.addTopic,
+    (state) => ({ isLoading: true, errorMessage: undefined }),
+  ),
+  on(
+    MatrixActions.getMatrixDataFailed,
+    MatrixActions.updateTaskFailed,
+    MatrixActions.deleteTaskFailed,
+    MatrixActions.updateTopicFailed,
+    MatrixActions.addTopicFailed,
+    (state, { message }) => ({ isLoading: false, errorMessage: message }),
+  ),
+  on(
+    MatrixActions.getMatrixDataSuccess,
+    MatrixActions.updateTaskSuccess,
+    MatrixActions.deleteTaskSuccess,
+    MatrixActions.addTopicSuccess,
+    MatrixActions.updateTopicSuccess,
+    (state) => ({ isLoading: false, errorMessage: undefined }),
+  ),
+);
