@@ -1,16 +1,9 @@
 import { Injectable, OnDestroy } from '@angular/core';
 import { select, Store } from '@ngrx/store';
-import { Observable, Subject } from 'rxjs';
+import { Observable, of, Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
-import { AppState } from '../store/app.state';
-import {
-  DeleteTask,
-  GetMatrixData,
-  SelectTask,
-  ToggleTopicVisibility,
-  UpdateTask,
-  UpdateTopic,
-} from './matrix.actions';
+import { AppState } from '../app.state';
+import { RequestStatus } from '../shared/request-status.interface';
 import {
   MatrixService as IMatrixService,
   Task,
@@ -18,13 +11,17 @@ import {
   Topic,
 } from './matrix.interfaces';
 import {
-  selectMatrixErrorMessage,
-  selectMatrixIsLoading,
+  selectCurrentTaskHistory,
+  selectMatrixActiveTasks,
+  selectMatrixActiveTasksByTopics,
+  selectMatrixDoneTasks,
+  selectMatrixDoneTasksByTopics,
+  selectMatrixRequestStatus,
+  selectMatrixTaskHistory,
   selectMatrixTasks,
   selectMatrixTasksByTopics,
   selectMatrixTopic,
   selectMatrixTopics,
-  selectTaskHistory,
 } from './matrix.selectors';
 
 @Injectable({
@@ -40,30 +37,6 @@ export class MatrixService implements IMatrixService, OnDestroy {
     this.unsubscribe$.complete();
   }
 
-  getData(): void {
-    this.store.dispatch(new GetMatrixData());
-  }
-
-  updateTask(task: Task): void {
-    this.store.dispatch(new UpdateTask(task));
-  }
-
-  deleteTask(taskid: number): void {
-    this.store.dispatch(new DeleteTask(taskid));
-  }
-
-  selectTask(task: Task): void {
-    console.log(task);
-    this.store.dispatch(new SelectTask(task.id));
-  }
-  toggleTopicVisibility(topicId: number): void {
-    this.store.dispatch(new ToggleTopicVisibility(topicId));
-  }
-
-  updateTopic(topic: Topic): void {
-    this.store.dispatch(new UpdateTopic(topic));
-  }
-
   selectTopics(): Observable<Topic[]> {
     return this.store.pipe(
       select(selectMatrixTopics),
@@ -71,16 +44,9 @@ export class MatrixService implements IMatrixService, OnDestroy {
     );
   }
 
-  selectIsLoading(): Observable<boolean> {
+  selectRequestStatus(): Observable<RequestStatus> {
     return this.store.pipe(
-      select(selectMatrixIsLoading),
-      takeUntil(this.unsubscribe$),
-    );
-  }
-
-  selectErrorMessage(): Observable<string | undefined> {
-    return this.store.pipe(
-      select(selectMatrixErrorMessage),
+      select(selectMatrixRequestStatus),
       takeUntil(this.unsubscribe$),
     );
   }
@@ -92,6 +58,20 @@ export class MatrixService implements IMatrixService, OnDestroy {
     );
   }
 
+  selectActiveTasks(): Observable<Task[]> {
+    return this.store.pipe(
+      select(selectMatrixActiveTasks),
+      takeUntil(this.unsubscribe$),
+    );
+  }
+
+  selectDoneTasks(): Observable<Task[]> {
+    return this.store.pipe(
+      select(selectMatrixDoneTasks),
+      takeUntil(this.unsubscribe$),
+    );
+  }
+
   selectTasksByTopics(): Observable<TaskDictionary> {
     return this.store.pipe(
       select(selectMatrixTasksByTopics),
@@ -99,9 +79,30 @@ export class MatrixService implements IMatrixService, OnDestroy {
     );
   }
 
-  selectTaskHistory(): Observable<Task[]> {
+  selectActiveTasksByTopics(): Observable<TaskDictionary> {
     return this.store.pipe(
-      select(selectTaskHistory),
+      select(selectMatrixActiveTasksByTopics),
+      takeUntil(this.unsubscribe$),
+    );
+  }
+
+  selectDoneTasksByTopics(): Observable<TaskDictionary> {
+    return this.store.pipe(
+      select(selectMatrixDoneTasksByTopics),
+      takeUntil(this.unsubscribe$),
+    );
+  }
+
+  selectCurrentTaskHistory(): Observable<Task[]> {
+    return this.store.pipe(
+      select(selectCurrentTaskHistory),
+      takeUntil(this.unsubscribe$),
+    );
+  }
+
+  selectMatrixTaskHistory(): Observable<number[]> {
+    return this.store.pipe(
+      select(selectMatrixTaskHistory),
       takeUntil(this.unsubscribe$),
     );
   }
@@ -111,5 +112,10 @@ export class MatrixService implements IMatrixService, OnDestroy {
       select(selectMatrixTopic, id),
       takeUntil(this.unsubscribe$),
     );
+  }
+
+  mockFunction(data: any): Observable<any> {
+    console.log(data);
+    return of(data);
   }
 }
