@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { TaskDictionary } from '../matrix.interfaces';
 import { TaskDotComponent } from './task-dot.component';
 
@@ -7,15 +7,17 @@ describe('TaskDotComponent', () => {
   let fixture: ComponentFixture<TaskDotComponent>;
   let taskDot: any;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [TaskDotComponent],
-    }).compileComponents();
+  beforeEach(
+    waitForAsync(() => {
+      TestBed.configureTestingModule({
+        declarations: [TaskDotComponent],
+      }).compileComponents();
 
-    fixture = TestBed.createComponent(TaskDotComponent);
-    component = fixture.componentInstance;
-    taskDot = fixture.debugElement.componentInstance;
-  }));
+      fixture = TestBed.createComponent(TaskDotComponent);
+      component = fixture.componentInstance;
+      taskDot = fixture.debugElement.componentInstance;
+    }),
+  );
 
   it('should create the task dot', () => {
     expect(taskDot).toBeTruthy();
